requirements.txt
Django==4.2.7
Pillow==10.1.0
requests==2.31.0
google-generativeai==0.7.2
python-dotenv==1.0.0
